librarian: rename docker
